fix(core): sign cookies with built-in jwt methods
